% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_color_cepesp}
\alias{scale_color_cepesp}
\alias{scale_fill_cepesp}
\title{CEPESP colour scales}
\usage{
scale_color_cepesp(palette = "principal", discrete = TRUE,
  reverse = FALSE, ...)

scale_fill_cepesp(palette = "principal", discrete = TRUE, reverse = FALSE,
  ...)
}
\description{
\code{scale_*_cepesp} functions are
}
