% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cepesp.R
\name{theme_cepesp}
\alias{theme_cepesp}
\title{CEPESP theme for ggplo2 graphs}
\usage{
theme_cepesp(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\description{
CEPESP theme for ggplo2 graphs
}
